Backend/requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.3.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Django==4.2.8
django-cors-headers==4.3.1
django-environ==0.12.0
django-filter==23.5
djangorestframework==3.14.0
filelock==3.20.1
fonttools==4.61.1
fsspec==2025.12.0
gunicorn==21.2.0
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
kombu==5.6.1
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==1.26.4
opencv-python==4.8.1.78
packaging==25.0
pandas==2.3.3
pillow==12.0.0
prompt_toolkit==3.0.52
psutil==7.1.3
psycopg2-binary==2.9.11
py-cpuinfo==9.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
redis==5.0.1
requests==2.31.0
scipy==1.16.3
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.5
sympy==1.14.0
thop==0.1.1.post2209072238
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.0.227
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
